app/packages/client/pyproject.toml
[project]
name = "radium226-vpn-passthrough-client"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = ["loguru", "pyyaml", "radium226-vpn-passthrough-messages", "radium226-vpn-passthrough-ipc"]

[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "radium226.vpn_passthrough.client"

[tool.uv.sources]
radium226-vpn-passthrough-messages = { workspace = true }
radium226-vpn-passthrough-ipc = { workspace = true }
radium226-vpn-passthrough-server = { workspace = true }

[dependency-groups]
dev = ["pytest>=9.0.2", "pytest-asyncio>=1.0.0", "ruff>=0.15.4", "ty>=0.0.19",
       "radium226-vpn-passthrough-server"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
src = ["src"]

[tool.ty.src]
include = ["src"]
